#11-9 CBV_SRV_UAVのディスクリプタヒープクラス追加 テクスチャマッピング実装完了

+++ RenderingEngine/src/DescriptorHeapRTV.cpp
@@ -19,12 +19,11 @@ HRESULT DescriptorHeapRTV::CreateDescriptorHeap(ID3D12Device& device)
 MYRESULT DescriptorHeapRTV::Create(ID3D12Device& device)
 {
 	// ハンドルのインクリメントサイズ取得
-	_handleIncrimentSize = device.GetDescriptorHandleIncrementSize(D3D12_DESCRIPTOR_HEAP_TYPE_RTV);
+	_handleIncrimentSize =
+		static_cast<SIZE_T>(device.GetDescriptorHandleIncrementSize(D3D12_DESCRIPTOR_HEAP_TYPE_RTV));
 
 	// ディスクリプタヒープ生成
-	if (FAILED(CreateDescriptorHeap(device))) {
-		return MYRESULT::FAILED;
-	}
+	if (FAILED(CreateDescriptorHeap(device))) { return MYRESULT::FAILED; }
 
 	return MYRESULT::SUCCESS;
 }
